# docs/swagger.yaml
basePath: /
definitions:
  domain.AssociateSensorRequest:
    properties:
      sensor_id:
        type: string
    required:
    - sensor_id
    type: object
  domain.ColumnSchemaType:
    properties:
      column:
        type: string
      type:
        enum:
        - float
        - int
        - bool
        - string
        type: string
    required:
    - column
    - type
    type: object
  domain.Device:
    properties:
      code:
        type: string
      created_at:
        type: string
      device_type:
        $ref: '#/definitions/domain.DeviceType'
      id:
        type: string
      room:
        $ref: '#/definitions/domain.Room'
    type: object
  domain.DeviceRequest:
    properties:
      code:
        maxLength: 50
        type: string
      device_type_id:
        type: string
      room_id:
        type: string
    required:
    - code
    - device_type_id
    type: object
  domain.DeviceType:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: string
      topic:
        type: string
    type: object
  domain.DeviceTypeRequest:
    properties:
      code:
        type: string
      description:
        type: string
      topic:
        maxLength: 50
        type: string
    required:
    - code
    - topic
    type: object
  domain.Room:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  domain.RoomRequest:
    properties:
      name:
        type: string
    type: object
  domain.SensorType:
    properties:
      code:
        type: string
      column_schema:
        additionalProperties:
          $ref: '#/definitions/domain.ColumnSchemaType'
        type: object
      description:
        type: string
      id:
        type: string
      payload_format:
        type: string
      qos_mqtt:
        type: integer
      readings_table_name:
        type: string
      topic:
        type: string
      value_mapping:
        type: object
    type: object
  domain.SensorTypeRequest:
    properties:
      code:
        type: string
      column_schema:
        additionalProperties:
          $ref: '#/definitions/domain.ColumnSchemaType'
        type: object
      description:
        type: string
      payload_format:
        enum:
        - json
        - raw
        type: string
      qos_mqtt:
        maximum: 2
        minimum: 0
        type: integer
      readings_table_name:
        type: string
      topic:
        maxLength: 50
        type: string
      value_mapping:
        type: object
    required:
    - code
    - column_schema
    - payload_format
    - readings_table_name
    - topic
    type: object
  domain.ValidationErrorResponse:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
    type: object
host: localhost:3030
info:
  contact: {}
  description: API for management of IoT system database
  title: IOT nonna core API
  version: "1.0"
paths:
  /device-types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DeviceType'
        "500":
          description: internal server error
          schema:
            type: string
      summary: Lists all device types
      tags:
      - deviceType
    post:
      consumes:
      - application/json
      parameters:
      - description: Device type payload
        in: body
        name: deviceType
        required: true
        schema:
          $ref: '#/definitions/domain.DeviceTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.DeviceType'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/domain.ValidationErrorResponse'
        "500":
          description: internal server error
          schema:
            type: string
      summary: Add device type
      tags:
      - deviceType
  /device-types/{id}:
    delete:
      parameters:
      - description: Device type id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: no content
        "404":
          description: device_type not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete device type
      tags:
      - deviceType
    get:
      parameters:
      - description: Device type id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DeviceType'
        "404":
          description: device_type not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Returns the device type based on the id
      tags:
      - deviceType
    put:
      consumes:
      - application/json
      parameters:
      - description: Device type id
        in: path
        name: id
        required: true
        type: string
      - description: Device type payload
        in: body
        name: deviceType
        required: true
        schema:
          $ref: '#/definitions/domain.DeviceTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.DeviceType'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/domain.ValidationErrorResponse'
        "404":
          description: device_type not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Update device type
      tags:
      - deviceType
  /devices:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Device'
            type: array
        "500":
          description: internal server error
          schema:
            type: string
      summary: Lists all devices
      tags:
      - devices
    post:
      consumes:
      - application/json
      parameters:
      - description: device payload
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/domain.DeviceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Device'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/domain.ValidationErrorResponse'
        "500":
          description: internal server error
          schema:
            type: string
      summary: Add a device
      tags:
      - devices
  /devices/{id}:
    delete:
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: no content
        "404":
          description: device not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete device
      tags:
      - devices
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Device'
        "404":
          description: device not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Returns the device with specified id
      tags:
      - devices
    put:
      consumes:
      - application/json
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      - description: device payload
        in: body
        name: device
        required: true
        schema:
          $ref: '#/definitions/domain.DeviceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Device'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/domain.ValidationErrorResponse'
        "404":
          description: device not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Update device based on id
      tags:
      - devices
  /devices/{id}/sensors:
    get:
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.SensorType'
            type: array
        "500":
          description: internal server error
          schema:
            type: string
      summary: Lists all sensors of selected device
      tags:
      - deviceSensors
    post:
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      - description: Associated sensor payload
        in: body
        name: sensorId
        required: true
        schema:
          $ref: '#/definitions/domain.AssociateSensorRequest'
      produces:
      - application/json
      responses:
        "204":
          description: status no content
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/domain.ValidationErrorResponse'
        "500":
          description: internal server error
          schema:
            type: string
      summary: Add sensor to device
      tags:
      - deviceSensors
  /devices/{id}/sensors/{sensorId}:
    delete:
      parameters:
      - description: device id
        in: path
        name: id
        required: true
        type: string
      - description: sesor id
        in: path
        name: sensorId
        required: true
        type: string
      responses:
        "204":
          description: status no content
        "404":
          description: sensors_devices not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete sensor from device
      tags:
      - deviceSensors
  /health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Db offline
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Check connection with db
      tags:
      - health
  /rooms:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Room'
            type: array
        "500":
          description: internal server error
          schema:
            type: string
      summary: lists all rooms
      tags:
      - rooms
    post:
      consumes:
      - application/json
      parameters:
      - description: Room payload
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/domain.RoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Room'
        "400":
          description: invalid request body
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create a new room
      tags:
      - rooms
  /rooms/{id}:
    delete:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: no content
        "404":
          description: room not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete the room
      tags:
      - rooms
    get:
      parameters:
      - description: Room ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Room'
        "404":
          description: room not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: returns the rooms with the specified id
      tags:
      - rooms
    put:
      consumes:
      - application/json
      parameters:
      - description: RoomId
        in: path
        name: id
        required: true
        type: string
      - description: Room payload
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/domain.RoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Room'
        "400":
          description: invalid request body
          schema:
            type: string
        "404":
          description: room not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Update room
      tags:
      - rooms
  /sensor-types:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.SensorType'
            type: array
        "500":
          description: internal server error
          schema:
            type: string
      summary: Lists all sensor types
      tags:
      - sensorTypes
    post:
      consumes:
      - application/json
      parameters:
      - description: Sensor type payload
        in: body
        name: sensorType
        required: true
        schema:
          $ref: '#/definitions/domain.SensorTypeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.SensorType'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/domain.ValidationErrorResponse'
        "500":
          description: internal server error
          schema:
            type: string
      summary: Add new sensor type
      tags:
      - sensorTypes
  /sensor-types/{id}:
    delete:
      parameters:
      - description: Sensor type id
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: no content
        "404":
          description: Sensor type not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete sensor type based on id
      tags:
      - sensorTypes
    get:
      parameters:
      - description: SensorType ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SensorType'
        "404":
          description: sensor_type not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Returns sensor type with specified id
      tags:
      - sensorTypes
    put:
      consumes:
      - application/json
      parameters:
      - description: sensor type id
        in: path
        name: id
        required: true
        type: string
      - description: sensor type payload
        in: body
        name: sensorType
        required: true
        schema:
          $ref: '#/definitions/domain.SensorTypeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SensorType'
        "400":
          description: validation error
          schema:
            $ref: '#/definitions/domain.ValidationErrorResponse'
        "404":
          description: sensor_type not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Update sensor type based on id
      tags:
      - sensorTypes
swagger: "2.0"